+++ Sources/StreamChatUI/MessageActionsPopup/ChatMessageActionsVC.swift
@@ -89,7 +89,9 @@ open class ChatMessageActionsVC: _ViewController, ThemeProvider {
             let currentUser = messageController.dataStore.currentUser(),
             let message = message,
             message.isDeleted == false
-        else { return [] }
+        else {
+            return []
+        }
 
         switch message.localState {
         case nil:
@@ -149,13 +151,17 @@ open class ChatMessageActionsVC: _ViewController, ThemeProvider {
                 editActionItem(),
                 deleteActionItem()
             ]
-        case .pendingSend, .pendingSync, .syncingFailed, .deletingFailed:
+        case .pendingSend,
+             .pendingSync,
+             .syncingFailed,
+             .deletingFailed,
+             .sending,
+             .syncing,
+             .deleting:
             return [
                 editActionItem(),
                 deleteActionItem()
             ]
-        case .sending, .syncing, .deleting:
-            return []
         }
     }